+++ CHANGELOG.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the master branch](https://github.com/gufolabs/gufo_acme/blob/master/CHANGELOG.md) guide.
 
-## [Unreleased]
-
 ## 0.1.1 - 2023-11-16
+
+### Fixed
 
 * Fixed `ACMEClient.from_state()` to return a proper subclass.
 * Fixed type annotation for `ACMEClient.__aenter__()` in subclasses.